Split AWS' handlers in multiple files
